8250: move SERIAL_PORT_DFNS into 8250_platform.c
There are currently three definitions of SERIAL_PORT_DFNS, and they
are all identical, aside from whitespace changes. The only user is
in drivers/tty/serial/8250/8250_platform.c.

Simplify this a little by moving the definitions into the place
where they are used, and use a Kconfig dependency to enable that
in place of the complex macros.

As Maciej Rozycki points out, the same code would be useful on any
ISA based machine, so instead of enabling this for 'M68K && ISA',
using all ISA machines is more complete, as this also includes
EISA based parisc and mips machines.

CONFIG_SERIAL_8250_DETECT_IRQ has no effect with that option and
can be hidden from users when SERIAL_8250_LEGACY is disabled or
unavailable.

Cc: "Maciej W. Rozycki" <macro@orcam.me.uk>
Signed-off-by: Arnd Bergmann <arnd@arndb.de>