serial: 8250: avoid serial8250_set_defaults() warning
On architectures without CONFIG_HAS_IOPORT, the 8250 console driver
started warning when the boot-time "nr_uarts" includes some early
uart definitions that are not initialized:

WARNING: CPU: 0 PID: 0 at drivers/tty/serial/8250/8250_port.c:470 serial8250_set_defaults+0x1a8/0x22c
Unsupported UART type 0

The problem here is that serial8250_set_defaults() gets called on a
port that has not been initialized, so the default port->iotype is the
implied 0 (UPIO_PORT), which is unavailable on architectures without
CONFIG_HAS_IOPORT. This is a preexisting bug that was introduced by
an earlier attempt to no longer crash when opening an uninitialized
uart port.

Between the many ways that a port can be added into this driver, it's
easy to get confused, so I tried to come up with the exact order that
things have to be initialized in:

 - serial8250_setup_port() must be called once for each port before
   anything else is done on it, so this needs to happen at the beginning
   of univ8250_console_init(), early_serial_setup() or serial8250_init(),
   whichever comes first.

 - univ8250_port_ops must be set up after the first call to
   serial8250_setup_port(), but before starting the 8250_rsa driver

 - the rest of __serial8250_isa_init_ports() must be called once on x86,
   alpha and m68k to get the initial ports, but only after
   univ8250_port_ops is set up. This can come from either
   serial8250_init() or from univ8250_console_init() if that is enabled.

 - serial8250_set_defaults() must be called once on each port added
   by __serial8250_isa_init_ports(), early_serial_setup()
   or serial8250_register_8250_port(), but only after the uart_port
   structure has been filled and before serial8250_isa_config() is called
   on that port.

Rearrange the code to move the handling of serial8250_setup_port() and
univ8250_port_ops back into the core, and do that for all ports regardless
of any runtime configuration, but leave the serial8250_set_defaults()
until the point at which a port is actually added. This adds a little
complexity at the moment, but works towards the point where this can
be simplified more easily in follow-up patches.

Fixes: 7c7e6c8924e7 ("tty: serial: handle HAS_IOPORT dependencies")
Fixes: f4c23a140d80 ("serial: 8250: fix null-ptr-deref in serial8250_start_tx()")
Reported-by: Guenter Roeck <linux@roeck-us.net>
Cc: Andy Shevchenko <andriy.shevchenko@linux.intel.com>
Cc: Maciej W. Rozycki <macro@orcam.me.uk>
Cc: Niklas Schnelle <schnelle@linux.ibm.com>
Cc: Jiri Slaby <jirislaby@kernel.org>
Closes: https://lore.kernel.org/all/38bf3b8b-81c4-42c9-902d-b5719f3881c5@roeck-us.net/
Signed-off-by: Arnd Bergmann <arnd@arndb.de>