serial: 8250: ensure port->dev is set during registration
A 8250 uart with port->dev set to NULL during early boot is treated
as an ISA port and gets assigned to the default 'serial8250_isa_devs'
platform device, which appears to make it work by accident on x86
machines, but not the way it was intended.

I found three drivers that get this wrong but also have a device
structure that can be meaningfully assigned here, so do this and
add a WARN_ON() in the registration to catch any future drivers
trying to do this.

The "mwave" modem driver also fails to set port->dev here, and
appears to do so in a way that is actively broken. It also looks
like configuring the uart is optional in that driver, so I'm no
touching that one here.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>