serial: 8250: decouple CONFIG_SERIAL_8250_SHARE_IRQ from platform driver
CONFIG_SERIAL_8250_SHARE_IRQ is mainly used to decide if the ISA-style PC
serial ports used shared interrupts or not. The implementation mixes this
with the default for plat_serial8250_port based UARTs that also get this
default despite having the information about the IRQ encoded in that data.

In particular, there are three platforms (fsl85xx, bcm2835, ioc3) that
attempt to enforce this behavior through Kconfig, but those don't actually
use plat_serial8250_port probing and instead use separate drivers that
call serial8250_register_8250_port() and ignore the Kconfig default.
In case of the devicetree based platforms, this is done by calling
uart_read_port_properties() to set the UPF_SHARE_IRQ flag.

Drop the Kconfig logic and the platform_device hack and instead rely on
the platforms setting the data correctly. Since there are only about two
dozen users of plat_serial8250_port, I could check them indivdually to
verify that they either set the flag or don't share interrupts. As it
turns out, the only ones that don't already set the flag correctly are
the five ISA card drivers that originally needed it.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>