[SUBMITTED 20250224] mm: unhide get_dump_page() function
The definition of get_dump_page() is guarded by CONFIG_ELF_CORE, but
the caller has now moved into a function that is built based on
CONFIG_COREDUMP, which leads to a possible link failure:

ld.lld-21: error: undefined symbol: get_dump_page
>>> referenced by coredump.c
>>>               fs/coredump.o:(dump_vma_snapshot) in archive vmlinux.a

Change the #ifdef block around the definition to match the caller.
In practice there is very little difference, as setting COREDUMP
but not ELF_CORE is not useful.

Fixes: ff41385709f0 ("coredump: Only sort VMAs when truncating or core_sort_vma sysctl is set")
Signed-off-by: Arnd Bergmann <arnd@arndb.de>