apne: convert to standalone driver
There is not much point pretending that this driver behaves like an
ISA card, when it is only ever used on a particular machine with a
specific adapter.

Simplify this by converting it from the 8290.ko library module
to including lib8390.c directly, like the other platform specific
variants. This requires a few mechanical changes

 - use ei_{in,out},{b,w}{_p,} consistently instead of the ISA PIO
   functions
 - add a custom net_device_ops structure
 - revive the gayle_{in,out}{b,w} helpers

Signed-off-by: Arnd Bergmann <arnd@arndb.de>