ne2000: split out ethernec driver
The ethernec adapter uses a custom bus interface and a platform device
to probe it, with the inb()/outb() helpers in m68k making it appear
just enough like an ISA device to use the ISA ne2000 driver.

To separate the asm/io.h functions from this hack, add a custom
wrapper for the 8390 ethernet driver for just ethernec. This is
a copy of the ne.c with a few changes:

 - the inb()/outb() helpers are replaced with direct calls to the
   isa_rom helpers, avoiding all the extra indirections.

 - the ISA bus probing including isapnp and Space.c hooks and
   ISA specific module parameters are all removed.

 - I/O is hardcoded to 8-bit mode instead of detecting
   16-bit mode based on the hardware platform

Signed-off-by: Arnd Bergmann <arnd@arndb.de>