m68k: drop ISA abstraction for atari/ethernec
The ethernec device is the only user of the rom isa mode, and this is
now converted to call into that directly.

NetUSBee/isp116x in theory also uses the ATARI_ROM_ISA helpers, but the
patch for it was never merged, and neither the upstream driver nor the
version posted used the inw/outw helpers.

With this gone, the inb/outb functions are only used for Q40 and no
longer need any runtime detection. The raw I/O functions are left
in place as they are used by the ethernec driver and could be used
by NetUSBee.

Link: https://lore.kernel.org/linux-m68k/1365208006-16227-13-git-send-email-schmitz@debian.org/
Signed-off-by: Arnd Bergmann <arnd@arndb.de>