x86: allow system rom to be mapped in /dev/mem
With CONFIG_STRICT_DEVMEM, the BIOS ROM area (0xf0000-0xfffff) is marked
as unusable by /dev/mem by default, except that a special hack on x86
used to still allow it. With that hack gone, a handful of applications that
try to read the DMI area are broken, so change the annotation to allow
it to be mapped again.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>