ARM: remove broken NO_IOPORT_MAP
A few platforms select NO_IOPORT_MAP, but this is broken in multiplatform
builds because this means drivers that require ioport_map() cannot be
used on other platforms.

The one platform that actually requires NO_IOPORT_MAP is RPC, and this
is not part of CONFIG_ARCH_MULTIPLATFORM either.

Remove the broken ones and change the nommu ones to not advertise
CONFIG_HAS_IOPORT at all.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>