[SUBMITTED 20250710] RDMA/mlx5: fix linking with CONFIG_INFINIBAND_USER_ACCESS=n
The check for rdma_uattrs_has_raw_cap() is not possible if user
access is disabled:

ERROR: modpost: "rdma_uattrs_has_raw_cap" [drivers/infiniband/hw/mlx5/mlx5_ib.ko] undefined!

Limit the check to configurations that have the option enabled
and instead assume the capability is not there otherwise.

From what I can tell, the UVERBS code in fs.c is not actually called
when INFINIBAND_USER_ACCESS is turned off, so this haz no effect
other than fixing the link error. A better change might be to not
build the code at all in that configuration, but I did not see
an obvious way to do that.

Fixes: 95a89ec304c3 ("RDMA/mlx5: Check CAP_NET_RAW in user namespace for flow create")
Signed-off-by: Arnd Bergmann <arnd@arndb.de>
---
If there is a better way of addressing the link failure, please just
treat this as a bug report