leds: gpio: make legacy gpiolib interface optional
There are still a handful of ancient mips/armv5/sh boards that use the
gpio_led:gpio member to pass an old-style gpio number, but all modern
users have been converted to gpio descriptors.

Make the code that deals with this optional so the legacy interfaces
can be left out for all normal builds.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>