nfc: marvell: convert to gpio descriptors
The only reason this driver seems to still use the legacy gpio
numbers is for the module parameter that lets users pass a different
reset gpio.

Since the fixed numbers are on their way out, and none of the platforms
this driver is used on would have them any more, remove the module
parameter and instead just use the reset information from firmware.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>