[v2] nfc: marvell: convert to gpio descriptors
The only reason this driver seems to still use the legacy gpio
numbers is for the module parameter that lets users pass a different
reset gpio.

Since the fixed numbers are on their way out, and none of the platforms
this driver is used on would have them any more, remove the module
parameter and instead just use the reset information from firmware.

Reviewed-by: Andy Shevchenko <andriy.shevchenko@linux.intel.com>
Link: https://lore.kernel.org/all/aJyGrhvViwoK2MeN@smile.fi.intel.com/
Signed-off-by: Arnd Bergmann <arnd@arndb.de>
----
v2: - use proper "reset" name
    - use logical levels

Note that Andy and Dmitry disagreed on the use of the extra NULL
check. I kept it in as preferred by Dmitry, to keep
nfcmrvl_chip_reset() and nfcmrvl_chip_halt() consistent.