Kbuild: use nolibc for header check
The header check is currently blocked on CONFIG_CC_CAN_LINK, but
in practice it is enough to have a set of libc headers for the target
architecture.

The nolibc directory is almost all that is required here, so add
a few tiny stubs to have enough data for building the headers being
checked.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>