mips: inst.h: fix incorrect padding
The 16-bit instructions are defined in terms of 'unsigned' bitfields,
which means there is implicit padding when the bits only add up to
16 instead of 32.

It is fairly clear that these were meant to be 16-bit types, so change
them accordingly.

This change gets flagged as an ABI break by scripts/check-uapi.sh.

Fixes: cd574704ec06 ("MIPS: MIPS16e: Add instruction formats.")
Signed-off-by: Arnd Bergmann <arnd@arndb.de>