[SUBMITTED 20251204] amd-xgbe: add CONFIG_INET dependency
The AMD XGBE driver gained a build-time dependency on IPv4 support:

WARNING: unmet direct dependencies detected for NET_SELFTESTS
  Depends on [n]: NET [=y] && PHYLIB [=y] && INET [=n]
  Selected by [y]:
  - AMD_XGBE [=y] && NETDEVICES [=y] && ETHERNET [=y] && NET_VENDOR_AMD [=y] && (OF_ADDRESS [=y] || ACPI [=n] || PCI [=y]) && HAS_IOMEM [=y] && (X86 || ARM64 [=y] || COMPILE_TEST [=y]) && PTP_1588_CLOCK_OPTIONAL [=y]

The selftests are already enabled when both CONFIG_INET and CONFIG_PHYLIB
are in use, but cannot be selected if one of them is turned off.

Add the IPv4 dependency to ensure that the driver cannot be enabled
if selftests are unavailable.

Fixes: 862a64c83faf ("amd-xgbe: introduce support ethtool selftest")
Signed-off-by: Arnd Bergmann <arnd@arndb.de>